moved all to the graphql-client module
