+++ graphql-client-support/pom.xml
@@ -41,24 +41,6 @@
       <artifactId>mockito-core</artifactId>
       <scope>test</scope>
     </dependency>
-
-    <dependency>
-      <groupId>com.expediagroup.streamplatform</groupId>
-      <artifactId>stream-registry-graphql-client</artifactId>
-      <version>0.8.3-SNAPSHOT</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.squareup.okhttp3</groupId>
-      <artifactId>okhttp</artifactId>
-      <version>4.2.2</version>
-    </dependency>
-    <dependency>
-      <groupId>com.expediagroup.streamplatform</groupId>
-      <artifactId>stream-registry-graphql-client-reactor</artifactId>
-      <version>0.8.3-SNAPSHOT</version>
-      <scope>compile</scope>
-    </dependency>
   </dependencies>
 
 </project>
