+++ it/pom.xml
@@ -29,18 +29,6 @@
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>com.expediagroup.streamplatform</groupId>
-      <artifactId>stream-registry-graphql-client-reactor</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.expediagroup.streamplatform</groupId>
-      <artifactId>stream-registry-graphql-client-support</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
 
     <dependency>
       <groupId>junit</groupId>
